+++ dbms/src/Storages/StorageTinyLog.cpp
@@ -388,7 +388,7 @@ BlockInputStreams StorageTinyLog::read(
     const SelectQueryInfo & /*query_info*/,
     const Context & context,
     QueryProcessingStage::Enum /*processed_stage*/,
-    const size_t max_block_size,
+    const UInt64 max_block_size,
     const unsigned /*num_streams*/)
 {
     check(column_names);
